Remaking service hierarchy.
